api/requirements.txt
fastapi
python-dotenv
pydantic
scikit-learn
joblib
numpy
deep-translator
uvicorn